+++ src/main/java/com/atlauncher/gui/dialogs/FileTypeDialog.java
@@ -1,0 +1,107 @@
+/*
+ * ATLauncher - https://github.com/ATLauncher/ATLauncher
+ * Copyright (C) 2013-2019 ATLauncher
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.atlauncher.gui.dialogs;
+
+import java.awt.BorderLayout;
+import java.awt.FlowLayout;
+import java.awt.GridBagConstraints;
+import java.awt.GridBagLayout;
+import java.awt.event.WindowAdapter;
+import java.awt.event.WindowEvent;
+
+import javax.swing.JButton;
+import javax.swing.JComboBox;
+import javax.swing.JDialog;
+import javax.swing.JFrame;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+
+import com.atlauncher.App;
+
+@SuppressWarnings("serial")
+public class FileTypeDialog extends JDialog {
+    private JPanel top;
+    private JPanel middle;
+    private JPanel bottom;
+
+    private JLabel selectorLabel;
+    private JComboBox<String> selector;
+
+    private JButton bottomButton;
+
+    public FileTypeDialog(String title, String labelName, String bottomText, String selectorText, String[] subOptions) {
+        super(App.settings.getParent(), title, ModalityType.APPLICATION_MODAL);
+        setSize(400, 175);
+        setLocationRelativeTo(null);
+        setLayout(new BorderLayout());
+        setDefaultCloseOperation(JFrame.DO_NOTHING_ON_CLOSE);
+        setResizable(false);
+
+        // Top Panel Stuff
+        top = new JPanel();
+        top.add(new JLabel(labelName));
+
+        // Middle Panel Stuff
+        middle = new JPanel();
+        middle.setLayout(new GridBagLayout());
+        GridBagConstraints gbc = new GridBagConstraints();
+
+        gbc.gridx = 0;
+        gbc.gridy = 0;
+        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
+        selectorLabel = new JLabel(selectorText + ": ");
+        middle.add(selectorLabel, gbc);
+
+        gbc.gridx++;
+        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
+        selector = new JComboBox<>();
+        for (String item : subOptions) {
+            selector.addItem(item);
+        }
+        middle.add(selector, gbc);
+
+        // Bottom Panel Stuff
+        bottom = new JPanel();
+        bottom.setLayout(new FlowLayout());
+        bottomButton = new JButton(bottomText);
+        bottomButton.addActionListener(e -> close());
+        bottom.add(bottomButton);
+
+        add(top, BorderLayout.NORTH);
+        add(middle, BorderLayout.CENTER);
+        add(bottom, BorderLayout.SOUTH);
+
+        addWindowListener(new WindowAdapter() {
+            public void windowClosing(WindowEvent arg0) {
+                close();
+            }
+        });
+
+        setVisible(true);
+    }
+
+    private void close() {
+        setVisible(false);
+        dispose();
+    }
+
+    public String getSelectorValue() {
+        return (String) this.selector.getSelectedItem();
+    }
+
+}